+++ walk/git_test.go
@@ -63,8 +63,8 @@ func TestGitReader(t *testing.T) {
 	}
 
 	as.Equal(32, count)
-	as.Equal(int32(32), statz.Value(stats.Traversed))
-	as.Equal(int32(0), statz.Value(stats.Matched))
-	as.Equal(int32(0), statz.Value(stats.Formatted))
-	as.Equal(int32(0), statz.Value(stats.Changed))
+	as.Equal(32, statz.Value(stats.Traversed))
+	as.Equal(0, statz.Value(stats.Matched))
+	as.Equal(0, statz.Value(stats.Formatted))
+	as.Equal(0, statz.Value(stats.Changed))
 }
